refs #731 - removed circular dependency in data/store.ts - added "ApolloClient" to the named exports to make it compatible with Angular2 AOT compile - moved dev @types to devDependencies otherwise they potentially brake projects that are importing apollo-client
