+++ src/index.ts
@@ -78,6 +78,7 @@ export {
   MutationQueryReducersMap,
   Subscription,
   ApolloStore,
+  ApolloClient
 };
 
 export default ApolloClient;
