+++ src/optimistic-data/store.ts
@@ -7,8 +7,11 @@ import {
 
 import {
   data,
-  NormalizedCache,
 } from '../data/store';
+
+import {
+  NormalizedCache,
+} from '../data/storeUtils';
 
 import {
   getDataWithOptimisticResults,
